fix: addressing reviews s4 v3.14: format
